+++ dag/education.yml
@@ -1,10 +1,12 @@
 steps:
   # World Bank EdStats
-  data://meadow/wb/2023-07-10/worldbank_education:
-  - snapshot://wb/2023-07-10/worldbank_pisa.csv
-  - snapshot://wb/2023-07-10/worldbank_literacy.csv
-  - snapshot://wb/2023-07-10/worldbank_pre_primary.csv
-  - snapshot://wb/2023-07-10/worldbank_primary.csv
-  - snapshot://wb/2023-07-10/worldbank_secondary.csv
-  - snapshot://wb/2023-07-10/worldbank_tertiary.csv
-  - snapshot://wb/2023-07-10/worldbank_expenditure_and_teachers.csv
+  data://meadow/wb/2023-07-10/education:
+  - snapshot://wb/2023-07-10/education_learning_outcomes.csv
+  - snapshot://wb/2023-07-10/education_pre_primary.csv
+
+  #- snapshot://wb/2023-07-10/education_literacy.csv
+  #- snapshot://wb/2023-07-10/education_pre_primary.csv
+  #- snapshot://wb/2023-07-10/education_primary.csv
+  #- snapshot://wb/2023-07-10/education_secondary.csv
+  #- snapshot://wb/2023-07-10/education_tertiary.csv
+  #- snapshot://wb/2023-07-10/education_expenditure_and_teachers.csv
