+++ snapshots/wb/2023-07-10/education_primary.csv.dvc
@@ -1,0 +1,14 @@
+meta:
+  name: 'World Bank Education Statistics: Primary Education'
+  publication_year: 2018
+  source_name: World Bank EdStats (2023)
+  source_published_by: 'World Bank Education Statistics: Primary Education, World
+    Bank EdStats (2023)'
+  url: https://databank.worldbank.org/reports.aspx?source=Education%20Statistics#
+  source_data_url:
+  license_url: https://www.worldbank.org/en/about/legal/terms-of-use-for-datasets
+  license_name: Creative Commons Attribution 4.0
+  date_accessed: 2023-07-10
+  is_public: true
+  description: >
+    World bank data on primary school education.
