+++ snapshots/wb/2023-07-10/education_secondary.csv.dvc
@@ -1,8 +1,8 @@
 meta:
-  name: 'World Bank Education Statistics: Primary Education'
+  name: 'World Bank Education Statistics: Secondary Education'
   publication_year: 2018
   source_name: World Bank EdStats (2023)
-  source_published_by: 'World Bank Education Statistics: Primary Education, World
+  source_published_by: 'World Bank Education Statistics: Secondary Education, World
     Bank EdStats (2023)'
   url: https://databank.worldbank.org/reports.aspx?source=Education%20Statistics#
   source_data_url:
@@ -11,9 +11,4 @@ meta:
   date_accessed: 2023-07-10
   is_public: true
   description: >
-    World bank data on primary school education.
-wdir: ../../../data/snapshots/worldbank_education/2023-07-10
-outs:
-- md5: 5f3f60487d8d4dd470ea12a24edd4ac5
-  size: 2821659
-  path: worldbank_primary.csv
+    World bank data on secondary school education.
