--- snapshots/wb/2023-07-10/worldbank_literacy.csv.dvc
@@ -1,21 +1,0 @@
-meta:
-  name: 'World Bank Education Statistics: Literacy'
-  publication_year: 2018
-  source_name: World Bank EdStats (2023)
-  source_published_by: 'World Bank Education Statistics: Literacy, World Bank EdStats
-    (2023)'
-  url: https://databank.worldbank.org/reports.aspx?source=Education%20Statistics#
-  source_data_url:
-  license_url: https://www.worldbank.org/en/about/legal/terms-of-use-for-datasets
-  license_name: Creative Commons Attribution 4.0
-  date_accessed: 2023-07-10
-  is_public: true
-  description: >
-    World bank data on literacy rates amongst adults (15+), youth (15-24), 25-64 year
-    olds, and eldery (65+) split by gender.
-
-wdir: ../../../data/snapshots/worldbank_education/2023-07-10
-outs:
-- md5: 86e2aee8a5d84abc1447d45090d9793c
-  size: 2591736
-  path: worldbank_literacy.csv
